materialize-motherduck: refactor config into a separate file
This is just moving all the config related stuff into a separate file, since it
will be getting more complicated soon.
